Radio Astronomy tutorials using the ipython notebook
A repository of the ipython notebooks used in Radio Astronomy
lectures.

First upload of Fourier Transform related notebooks.

Note:
These original notebooks were created with the v0.12 version of the
ipython notebook, to be compatible with the computer lab used
for the course. Notebooks are not forward but not backward compatible.
v2.x notebooks cannot be imported into v0.12.